refactoring simulations
